# api/api.yml
openapi: 3.0.3
info:
  title: CMS
  version: "0.1"
...
security:
  - BearerAuth: []
paths:
...
  /api/posts:
    post:
      tags:
        - posts
      summary: Créer un nouvel article
      description: Créer un article en fournissant un titre et un contenu
      security: 
      - BearerAuth: []
      operationId: createPost
      requestBody:
        description: Créer un nouvel article
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Titre de l'article
                content:
                  type: string
                  description: Contenu de l'article
                Picture:
                  type: string
                  format: binary
                  description: Image associée à l'article (facultatif)
                  nullable: true
                tags:
                  type: array
                  items:
                    type: string
                  description: Liste des tags associés à l'article
      responses:
        "201":
          description: Article créé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Requête invalide
        "422":
          description: Erreur de validation

    get:
      tags:
        - posts
      summary: Récupérer la liste des articles
      description: Récupérer la liste des articles filtrable par tags ou une chaine de recherche
      operationId: getPostsByTags
      parameters:
        - name: tags
          in: query
          description: Une list de tags pour filtrer les posts
          required: false
          schema:
            type: array
            items:
              type: string
        - name: query
          in: query
          description: une chaine de caractère pour filtrer les posts
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Liste des articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "404":
          description: Aucun article trouvé
...
components:
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        tags:
          type: array
          items:
            type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT